Made the backend the root directory, with the frontend being the "public" path
