--- backend/.gitignore
@@ -1,17 +1,0 @@
-# dormdesign specific
-.env
-dormdesign
-
-# general
-
-*.exe
-*.exe~
-*.dll
-*.so
-*.dylib
-
-# Test binary, built with `go test -c`
-*.test
-
-# Output of the go coverage tool, specifically when used with LiteIDE
-*.out
