--- backend/go.mod
@@ -1,16 +1,0 @@
-module github.com/gmisail/dormdesign
-
-go 1.15
-
-require (
-	github.com/davecgh/go-spew v1.1.1 // indirect
-	github.com/go-redis/redis/v8 v8.2.2
-	github.com/google/uuid v1.1.2
-	github.com/gorilla/websocket v1.4.2
-	github.com/joho/godotenv v1.3.0
-	github.com/kr/text v0.2.0 // indirect
-	github.com/labstack/echo/v4 v4.1.17
-	github.com/niemeyer/pretty v0.0.0-20200227124842-a10e7caefd8e // indirect
-	gopkg.in/check.v1 v1.0.0-20200227125254-8fa46927fb4f // indirect
-	gopkg.in/rethinkdb/rethinkdb-go.v6 v6.2.1
-)
